expression-measurement/streaming/python-streaming-example/pyproject.toml
[project]
name = "hume-stream-simple"
version = "0.1.0"
description = "Simple Hume streaming example using language model"
requires-python = ">=3.9"
dependencies = ["hume==0.13.8", "python-dotenv"]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main"]
